requirements.txt
amazon-textract-caller==0.2.3
amazon-textract-response-parser==1.0.2
amazon-textract-textractor==1.7.10
langchain==0.0.335
openai==1.26.0
python-dotenv==1.0.1
openpyxl==3.1.2
XlsxWriter==3.2.0
PyMuPDF==1.24.4
pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
streamlit==1.32.2
boto3
azure-identity==1.17.1
streamlit_pdf_viewer
streamlit_pdf_reader
fastapi
uvicorn
azure-cosmos==4.7.0
python-decouple==3.8
aioboto3==13.1.1
fuzzywuzzy
python-multipart
azure-ai-documentintelligence==1.0.0b2